update babel preset test
