--- packages/babel-preset/tests/fixtures/no-decorator.example.txt
@@ -1,20 +1,0 @@
-"use strict";
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.ExampleClass = void 0;
-
-class ExampleClass {
-  method() {
-    return 'bar';
-  }
-
-}
-
-exports.ExampleClass = ExampleClass;
-
-(() => {
-  const example = new ExampleClass();
-  console.log(example.method());
-})();
