--- packages/babel-preset/tests/fixtures/test-script.no-decorator.ts
@@ -1,10 +1,0 @@
-export class ExampleClass {
-  method() {
-    return 'bar';
-  }
-}
-
-(() => {
-  const example = new ExampleClass();
-  console.log(example.method());
-})();
