--- packages/babel-preset/tests/fixtures/tsconfig.json
@@ -1,18 +1,0 @@
-{
-    "extends": "@shopify/typescript-configs/library",
-    "compilerOptions": {
-      "baseUrl": ".",
-      "rootDir": ".",
-      "outDir": "./build",
-      "pretty": true,
-      "jsx": "react",
-      "strict": true,
-      "importHelpers": true,
-      "skipLibCheck": false,
-      "esModuleInterop": true,
-      "strictNullChecks": true
-    },
-    "include": ["./**/*"],
-    "exclude": ["**/*.test.ts", "**/*.test.tsx"]
-  }
-
